requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.2.0
torchvision==0.17.0
transformers==4.38.2
Pillow==10.1.0
numpy==1.26.2
pydantic>=2.7.4,<3.0.0
pandas==2.1.4
datasets==2.17.1
scikit-learn==1.4.1.post1
accelerate==0.27.2